Доробив множення ділення та віднімання

+++ calculator.h
@@ -5,6 +5,6 @@
 
 void dodavanie(int pervoechislo, int vtoroechislo);
 void mnoshenie(int pervoechislo, int vtoroechislo);
-void delenie(int pervoechislo, int vtoroechislo);
+void delenie(float pervoechislo, float vtoroechislo);
 void vidnimana(int pervoechislo, int vtoroechislo);
 void vvodenie_v_stepen(int chislo, int stepen);
